Fix storage size per institute
