+++ src/main/java/de/mpg/imeji/presentation/admin/StorageStatisticsBean.java
@@ -1,29 +1,109 @@
 package de.mpg.imeji.presentation.admin;
 
+import java.io.IOException;
+import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
+import java.util.concurrent.Future;
 
 import javax.faces.bean.ManagedBean;
+import javax.faces.bean.ViewScoped;
 
 import org.apache.commons.io.FileUtils;
+import org.apache.log4j.Logger;
 
+import de.mpg.imeji.logic.batch.StorageUsageAnalyseJob;
+import de.mpg.imeji.logic.config.Imeji;
+import de.mpg.imeji.logic.model.CollectionImeji;
+import de.mpg.imeji.logic.model.Item;
+import de.mpg.imeji.logic.model.User;
+import de.mpg.imeji.logic.search.Search;
+import de.mpg.imeji.logic.search.Search.SearchObjectTypes;
+import de.mpg.imeji.logic.search.factory.SearchFactory;
+import de.mpg.imeji.logic.search.factory.SearchFactory.SEARCH_IMPLEMENTATIONS;
+import de.mpg.imeji.logic.search.jenasearch.JenaCustomQueries;
+import de.mpg.imeji.logic.security.user.UserService;
 import de.mpg.imeji.logic.statistic.StatisticsService;
 
 @ManagedBean
+@ViewScoped
 public class StorageStatisticsBean {
-
-  private final List<Institute> institutes = new ArrayList<>();
-  private final String allFileSize;
+  private static final Logger LOGGER = Logger.getLogger(StorageStatisticsBean.class);
+  private List<Institute> institutes = new ArrayList<>();
+  private String allFileSize;
+  private String numberOfFilesInStorage;
+  private String sizeOfFilesinStorage;
+  private String freeSpaceInStorage;
+  private String lastUpdateStorageStatistics;
+  private Future<Integer> storageAnalyseStatus;
 
   public StorageStatisticsBean() {
-    final StatisticsService controller = new StatisticsService();
-    for (final String institute : controller.getAllInstitute()) {
-      institutes
-          .add(new Institute(institute, controller.getUsedStorageSizeForInstitute(institute)));
+    try {
+      final StatisticsService controller = new StatisticsService();
+      for (final String institute : controller.getAllInstitute()) {
+        institutes
+            .add(new Institute(institute, controller.getUsedStorageSizeForInstitute(institute)));
+      }
+      allFileSize = FileUtils.byteCountToDisplaySize(controller.getAllFileSize());
+      final StorageUsageAnalyseJob storageUsageAnalyse = new StorageUsageAnalyseJob();
+      this.numberOfFilesInStorage = Integer.toString(storageUsageAnalyse.getNumberOfFiles());
+      this.sizeOfFilesinStorage =
+          FileUtils.byteCountToDisplaySize(storageUsageAnalyse.getStorageUsed());
+      this.freeSpaceInStorage =
+          FileUtils.byteCountToDisplaySize(storageUsageAnalyse.getFreeSpace());
+      this.lastUpdateStorageStatistics = storageUsageAnalyse.getLastUpdate();
+    } catch (IOException | URISyntaxException e) {
+      LOGGER.error("Error constructing StorageUsageAnalyseJob", e);
     }
-    allFileSize = FileUtils.byteCountToDisplaySize(controller.getAllFileSize());
+  }
+
+  /**
+   * Start the job {@link StorageUsageAnalyseJob}
+   *
+   * @throws IOException
+   * @throws URISyntaxException
+   */
+  public String analyseStorageUsage() throws IOException, URISyntaxException {
+    storageAnalyseStatus = Imeji.getEXECUTOR().submit(new StorageUsageAnalyseJob());
+    return "";
+  }
+
+  /**
+   * return count of all {@link CollectionImeji}
+   *
+   * @return
+   */
+  public int getAllCollectionsSize() {
+    final Search search =
+        SearchFactory.create(SearchObjectTypes.COLLECTION, SEARCH_IMPLEMENTATIONS.JENA);
+    return search.searchString(JenaCustomQueries.selectCollectionAll(), null, null, 0, -1)
+        .getNumberOfRecords();
+  }
+
+  /**
+   * return count of all {@link User}
+   *
+   * @return
+   */
+  public int getAllUsersSize() {
+    try {
+      return new UserService().searchUserByName("").size();
+    } catch (final Exception e) {
+      return 0;
+    }
+  }
+
+  /**
+   * return count of all {@link Item}
+   *
+   * @return
+   */
+  public int getAllImagesSize() {
+    final Search search = SearchFactory.create(SearchObjectTypes.ITEM, SEARCH_IMPLEMENTATIONS.JENA);
+    return search.searchString(JenaCustomQueries.selectItemAll(), null, null, 0, -1)
+        .getNumberOfRecords();
   }
 
   public ArrayList<Institute> getInstitutes() {
@@ -43,6 +123,29 @@ public class StorageStatisticsBean {
 
   public String getAllFileSize() {
     return allFileSize;
+  }
+
+  public String getNumberOfFilesInStorage() {
+    return numberOfFilesInStorage;
+  }
+
+  public String getSizeOfFilesinStorage() {
+    return sizeOfFilesinStorage;
+  }
+
+  public String getFreeSpaceInStorage() {
+    return freeSpaceInStorage;
+  }
+
+  public String getLastUpdateStorageStatistics() {
+    return lastUpdateStorageStatistics;
+  }
+
+  public boolean getStorageAnalyseStatus() {
+    if (storageAnalyseStatus != null) {
+      return storageAnalyseStatus.isDone();
+    }
+    return true;
   }
 
 }
