+++ src/main/webapp/AdminOverview.xhtml
@@ -130,9 +130,6 @@
 										<a href="#{Navigation.applicationUrl}admin/statistics" title="#{lbl.statistics}">#{lbl.statistics}</a>
 									</li>
 									<li>
-										<a href="#{Navigation.applicationUrl}admin/storageusage" title="#{lbl.storage_usage}">#{lbl.usage}</a>
-									</li>
-									<li>
 										<a href="#{Navigation.applicationUrl}admin/tools" title="#{lbl.tools}">#{lbl.tools}</a>
 									</li>
 								</ul>
