--- src/main/webapp/AdminStorageUsage.xhtml
@@ -1,57 +1,0 @@
-<ui:composition template="templates/template_site.xhtml"
-	xmlns="http://www.w3.org/1999/xhtml"
-	xmlns:ui="http://java.sun.com/jsf/facelets"
-	xmlns:h="http://java.sun.com/jsf/html"
-	xmlns:f="http://java.sun.com/jsf/core"
-	xmlns:menu="http://java.sun.com/jsf/composite/components/menu">
-	<ui:param name="headline" value="#{lbl.storage_space_per_institute}" />
-	<ui:param name="currentMainMenu" value="admin" />
-	<ui:param name="backLink1" value="#{Navigation.adminUrl}"/>
-	<ui:param name="backlinkTitle1" value="Admin"/>
-	<ui:define name="actionsMenu">
-
-	</ui:define>
-
-	<ui:define name="mainContent">
-
-		<!-- Are You Admin? -->
-		<h:panelGroup layout="block" styleClass="imj_pageMessageArea" rendered="#{!Auth.sysAdmin}">
-			<h:outputText value="#{lbl.view_page_disallowed}" styleClass="imj_messageInfo" rendered="#{!Auth.loggedIn}" />
-			<h:outputText value="#{lbl.view_page_disallowed2}" styleClass="imj_messageInfo" rendered="#{Auth.loggedIn}" />
-		</h:panelGroup>
-
-		<ui:fragment rendered="#{Auth.sysAdmin}">
-			<!-- Only for Admins -->
-			<!-- Admin - TiledList -->
-			<div class="imj_administrationTiledList imj_maxWidth">
-				<div class="imj_listBody">
-					<h:panelGroup layout="block" styleClass="imj_adminPanel imj_storageInformation" rendered="#{Auth.sysAdmin}">
-						<div class="imj_content">
-							<div class="imj_admindataSet">
-								<div class="imj_admindataLabel" title="#{lbl.institute}">#{lbl.all}</div>
-								<div class="imj_admindataValue">
-									<div class="imj_admindataValueEntry">#{storageStatisticsBean.allFileSize}</div>
-								</div>
-							</div>
-							<!-- Table -->
-							<ui:repeat var="institute"
-								value="#{storageStatisticsBean.institutes}">
-								<div class="imj_admindataSet">
-									<div class="imj_admindataLabel" title="#{lbl.institute}">#{institute.name}</div>
-									<div class="imj_admindataValue">
-										<div class="imj_admindataValueEntry">#{institute.storageString}</div>
-									</div>
-								</div>
-							</ui:repeat>
-						</div>
-						<!-- END CONTENT -->
-					</h:panelGroup>
-				</div>
-				<!-- END LIST BODY -->
-			</div>
-			<!-- END ADMIN LIST -->
-		</ui:fragment>
-
-	</ui:define>
-
-</ui:composition>
